TADL-specific barcode transforms for SIP2
Changes behavior when looking up patrons by barcode:

1. Only search for active cards -- slightly different approach
than is used to limit to active cards in master today.

2. If no active cards are found matching the requested barcode,
and the barcode begins with a letter, lowercase the barcode and
trim a variable number of characters from the end of the barcode.

Any barcode that worked before will work after this change, with
the exception of cards that are marked "inactive".